backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic
pydantic-settings
alembic
passlib
bcrypt<4.0
python-jose[cryptography]
pydantic[email]
python-multipart
cryptography
apscheduler
redis
websockets
PyJWT
docker